src/app/student-discipline-component/student-discipline.component.ts
import { Component, OnInit } from '@angular/core';
import {FormBuilder, FormGroup} from "@angular/forms";
import {Student} from "../student";
import {StudentDisciplineService} from "../student-discipline.service";

@Component({
  selector: 'app-student-discipline',
  templateUrl: './student-discipline.component.html',
  styleUrls: ['./student-discipline.component.css']
})
export class StudentDisciplineComponent implements OnInit {
  students: Student[];
  form: FormGroup = this.formBuilder.group({
    student: [''],
    violation: [''],
    punishment: [''],
  })

  constructor(private formBuilder: FormBuilder, private studentDisciplineService: StudentDisciplineService) {
  }

  ngOnInit(): void {
    this.studentDisciplineService.getStudents().subscribe(response => {

      //if response has an error
      //displayError(response.error)

      // else
      this.students = response;
    })

    //call the student service
    //save the response in a students propety
  }

}
